% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncatetext.R
\name{renderTruncatetext}
\alias{renderTruncatetext}
\title{Widget render function for use in Shiny}
\usage{
renderTruncatetext(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

