% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncatetext.R
\name{truncatetext}
\alias{truncatetext}
\title{Truncated Text}
\usage{
truncatetext(text, lines = 2, duration = 400, more = NULL, less = NULL,
  width = "100\%", height = "auto")
}
\arguments{
\item{text}{text inside the box}

\item{lines}{lines to show}

\item{duration}{determining how long the animation will run, in milliseconds}

\item{more}{label for the more button}

\item{less}{label for the less button}
}
\description{
Makes a truncatable text element.
}

